# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: JadiSah API
  version: 0.1.0
  description: |
    Kontrak API resmi JadiSah (Wedding Planner). File ini adalah **Single Source of Truth**
    untuk Backend (Go, via oapi-codegen), Frontend (TypeScript generated), dan pengujian (Bruno).

    ## Autentikasi
    Autentikasi ditangani oleh [Limen](https://limenauth.dev). Setelah sign-in / sign-up,
    server mengirim cookie `limen_session` (HttpOnly). Klien non-browser (Bruno, mobile)
    dapat memakai header `Authorization: Bearer <token>` dengan token dari response header
    `Set-Auth-Token`.

    ## Otorisasi (IAM / RBAC)
    Setiap operasi dilindungi oleh permission `resource:action`. Role bawaan:
    `SUPER_ADMIN`, `ADMIN`, `USER`. Fitur tertentu juga dibatasi oleh `subscription_tier`
    (`FREE`, `PREMIUM`, `VIP`).

    > Endpoint dengan tag `auth` diimplementasikan oleh Limen (bukan oleh kode hasil generate),
    > namun tetap didokumentasikan di sini agar kontraknya jelas untuk Frontend.
servers:
  - url: /
    description: Server yang sama (di belakang Traefik Ingress)
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /api/users/me:
    get:
      operationId: getMyProfile
      tags: [users]
      summary: Profil lengkap user login (role, permission, subscription, pasangan)
      responses:
        "200":
          description: Profil user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "401": { $ref: "#/components/responses/Unauthorized" }
    patch:
      operationId: updateMyProfile
      tags: [users]
      summary: Ubah nama user login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profil berhasil diubah
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }

  # ──────────────────────────────────────────────────────────────────────────
  # COUPLES
  # ──────────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: limen_session
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
    BadRequest:
      description: Request tidak valid
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Belum login / session tidak valid
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Error:
      type: object
      required: [message]
      properties:
        message: { type: string, example: "forbidden" }
        code: { type: string, example: "FORBIDDEN" }
...
    RoleName:
      type: string
      enum: [SUPER_ADMIN, ADMIN, USER]
    SubscriptionTier:
      type: string
      enum: [FREE, PREMIUM, VIP]

    # ── Auth (Limen) ────────────────────────────────────────────────────────
...
    UserSummary:
      type: object
      required: [id, email]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        first_name: { type: string, nullable: true }
        last_name: { type: string, nullable: true }
    Profile:
      type: object
      required: [id, email, subscription_tier, roles, permissions, created_at]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        first_name: { type: string, nullable: true }
        last_name: { type: string, nullable: true }
        email_verified: { type: boolean }
        subscription_tier: { $ref: "#/components/schemas/SubscriptionTier" }
        subscription_expires_at: { type: string, format: date-time, nullable: true }
        roles:
          type: array
          items: { $ref: "#/components/schemas/RoleName" }
        permissions:
          type: array
          description: Daftar permission efektif dalam format `resource:action`
          items: { type: string, example: "couple:invite" }
        partner:
          allOf:
            - $ref: "#/components/schemas/UserSummary"
          nullable: true
          description: Pasangan yang tertaut (null jika belum)
        created_at: { type: string, format: date-time }
    UpdateProfileRequest:
      type: object
      minProperties: 1
      properties:
        first_name: { type: string, minLength: 1, maxLength: 100 }
        last_name: { type: string, minLength: 1, maxLength: 100 }

    # ── Couples ─────────────────────────────────────────────────────────────
...